+++ lib/app.dart
@@ -1,8 +1,6 @@
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 
-import 'src/presentation/screens.dart';
-
 class App extends StatelessWidget {
   const App({Key? key}) : super(key: key);
