+++ lib/data/src/repository/health_management.dart
@@ -1,14 +1,13 @@
-import 'package:flutter/material.dart';
-
-import '../provider/provider.dart';
-import 'entity.dart';
-
-/// Damage delear:on bullet, ship
-abstract class IShipHealth {
-  double health();
-}
 
 // enemyShip Bullet Damage
+import 'package:flutter/material.dart';
+import 'package:space_craft/domain/src/entities/game_object.dart';
+import 'package:space_craft/domain/src/entities/health_management.dart';
+import 'package:space_craft/domain/src/entities/vector2.dart';
+import 'package:space_craft/presentation/src/core/providers/object_scalar.dart';
+
+import 'vector2.dart';
+
 class DamageOnEB implements IShipHealth {
   IShipHealth iShipHealth;
