+++ lib/data/src/util/constants/enums.dart
