+++ lib/domain/src/entities/bullet.dart
@@ -1,0 +1,5 @@
+import 'game_object.dart';
+
+abstract class IBullet implements GameObject {
+  int? get id;
+}
