+++ lib/domain/src/entities/entity.dart
@@ -1,2 +1,4 @@
 // entity package contains all the data classes that are used by the entire application.
 // We can encounter the usage of these entities in any module
+
+
