+++ lib/domain/src/entities/health_management.dart
@@ -1,0 +1,7 @@
+import 'package:flutter/material.dart';
+ 
+
+/// Damage delear:on bullet, ship
+abstract class IShipHealth {
+  double health();
+}
