+++ lib/domain/src/entities/score.dart
@@ -1,0 +1,3 @@
+abstract class IPlayerScore {
+  int score();
+}
