+++ lib/domain/src/entities/setting.dart
