+++ lib/domain/src/entities/vector2.dart
@@ -1,0 +1,3 @@
+
+
+ // todo create vector abstraction
