+++ lib/presentation/src/feature/on_play/utils/utils.dart
@@ -1,4 +1,4 @@
-export 'enemy_hanlader.dart';
+export '../../../../../data/src/helper/enemy_hanlader.dart';
 export 'object_collision_checker.dart';
 export 'player_movement_handler.dart';
 export 'player_ship_paint.dart';
